refactoring AttributeComparator for better reuse
