+++ service/src/main/java/org/databiosphere/workspacedataservice/shared/model/RecordAttributes.java
@@ -38,20 +38,6 @@ public class RecordAttributes {
 		this.attributes.putAll(attributes);
 	}
 
-	private record AttributeComparator(String primaryKey) implements Comparator<String> {
-
-		@Override
-		public int compare(String o1, String o2) {
-			if (o1.equals(primaryKey)) {
-				return -1;
-			}
-			if (o2.equals(primaryKey)) {
-				return 1;
-			}
-			Collator collator = Collator.getInstance();
-			return collator.compare(o1, o2);
-		}
-	}
 
 	/**
 	 * creates a RecordAttributes with no keys/values
